+++ src/assets/sass/section/_nav.scss
@@ -1,6 +1,6 @@
 .navbar {
   position: fixed;
-  bottom: 0;
+  bottom: -1%;
   left: 50%;
   transform: translateX(-50%);
   width: 100%;
